+++ P2Project/P2Project/ViewModel/MakeExerciseViewModel.cs
@@ -129,7 +129,7 @@ namespace P2Project.ViewModel
         private void ExerciseCreateClick(object param)
         {
             //TRY
-            float sum = ExerciseProfile.Auditory + ExerciseProfile.Kinesthetic + ExerciseProfile.Verbal + ExerciseProfile.Visual;
+            decimal sum = ExerciseProfile.Auditory + ExerciseProfile.Kinesthetic + ExerciseProfile.Verbal + ExerciseProfile.Visual;
             ExerciseProfile.Auditory /= sum;
             ExerciseProfile.Kinesthetic /= sum;
             ExerciseProfile.Verbal /= sum;
